moved Waiver logic to its own component and cleared signature errors on stroke end
